Added new CSS for login, search, and submit buttons
