+++ src/components/Playlist/Playlist.module.css
@@ -4,10 +4,6 @@
   text-align: center;
 }
 
-.SaveButton {
-  padding: 0.5em;
-}
-
 .PlaylistName {
   display: flex;
   border-style: solid;
