+++ src/components/SearchBar/SearchBar.module.css
@@ -1,7 +1,17 @@
 .SearchBar {
-  margin-left: 7%;
   margin-top: 5px;
   margin-bottom: 5px;
-  width: 40%;
-  text-align: center;
+  padding-left: 30%;
+  padding-right: 30%;
+}
+
+form {
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+}
+
+input {
+  width: 70%;
+  height: 2em;
 }
