+++ src/containers/LoginContainer/LoginContainer.module.css
@@ -6,4 +6,10 @@ p {
 
 .login {
   font-weight: bold;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: space-evenly;
+  color: white;
+  padding: 10px;
 }
